Extract debug utilities to shared module
Created packages/core/web/debug.ts with reusable debug tools:
- createDebugTools(ogma, control) - returns debug API
- toggleCommentDebug() - toggle comment bounds visualization
- isCommentDebugEnabled() - check if enabled
- destroy() - cleanup all debug layers

Features:
- Press 'D' to toggle comment debug
- Automatic refresh on view/annotation changes
- Proper cleanup on destroy

Usage in any playground:
  import { createDebugTools } from "./debug";
  const debug = createDebugTools(ogma, control);
